+++ package/lib/src/core/view/default_page_configuration.dart
